requirements.txt
pvporcupine
pyaudio
numpy
sounddevice
webrtcvad
openai-whisper
pyttsx3
google-generativeai
python-dotenv
